# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DeguBase – workspace and table management API.",
        "title": "DeguBase API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/workspaces/{wsCode}/tables/{tableCode}/rows/{rowID}/graph": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Traverses all row-link columns reachable from the given row up to the\nspecified depth and returns a graph with nodes and directed edges.\nCyclic references (A→B→C→A) are handled safely — each node appears exactly once.\nOrphaned links (the referenced row was deleted) produce a ghost node so the edge remains visible.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rows"
                ],
                "summary": "Build relationship graph for a row",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "table code",
                        "name": "tableCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "row ID",
                        "name": "rowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "max traversal depth (default 5, max 10)",
                        "name": "depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/records.GraphResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/workspaces/{wsCode}/tables/{tableCode}/rows/{rowID}/referencing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated records from any table in the workspace that\nlink to the given row via a row-link column.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rows"
                ],
                "summary": "List referencing records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace code",
                        "name": "wsCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "table code",
                        "name": "tableCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "row ID",
                        "name": "rowID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PagedReferencingRows"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "type": "string",
                    "example": "not found"
                }
            }
        },
...
        "models.PagedReferencingRows": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReferencingRow"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "models.ReferencingRow": {
            "type": "object",
            "properties": {
                "sourceRowId": {
                    "type": "integer"
                },
                "sourceRowLabel": {
                    "type": "string"
                },
                "sourceTableCode": {
                    "type": "string"
                },
                "sourceTableName": {
                    "type": "string"
                },
                "viaColumnCode": {
                    "type": "string"
                },
                "viaColumnName": {
                    "type": "string"
                }
            }
        },
...
        "records.GraphEdge": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "via_column": {
                    "description": "ViaColumn is the row-link column code on the source row.",
                    "type": "string"
                },
                "via_column_name": {
                    "type": "string"
                }
            }
        },
        "records.GraphNode": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data holds the code-keyed cell data. Null for orphaned (deleted) rows.",
                    "type": "object"
                },
                "id": {
                    "description": "ID uniquely identifies the node as \"{tableCode}:{rowID}\".",
                    "type": "string"
                },
                "label": {
                    "description": "Label is the best human-readable name for the row (first text column, or \"Row #N\").\nDeleted/orphaned rows carry their stale label with a \" (deleted)\" deleted suffix.",
                    "type": "string"
                },
                "row_id": {
                    "type": "integer"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "records.GraphResponse": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/records.GraphEdge"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/records.GraphNode"
                    }
                },
                "root": {
                    "description": "Root is the node ID of the starting record.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}